chore: 모듈화
- 상수 객체, 게임 멘트 출력 메서드, 입력 유효성 검사 메서드
